fix: address PR #63 review findings
- Pico HAL: promote gpio_input_configured_mask to uint64_t and shift with
  1ull so GPIO 32-47 on RP2350 (pico2_w, the default board) no longer hit
  C/C++ undefined behavior after passing is_valid_gpio().
- CLI: declare @devicesdk/core as a workspace:* runtime dep so the init
  template's resolvePackageVersion returns the installed version in
  published usage; also resolves a latent runtime-import bug in env.ts
  which imports ENV_VAR_KEY_REGEX at runtime.
- Core: expose "./package.json" in the package exports so
  createRequire(...).resolve("@devicesdk/core/package.json") works under
  Node's exports-enforced resolution (previously blocked).
- CLI: drop hardcoded DEFAULT_CORE_VERSION / DEFAULT_CLI_VERSION fallbacks
  in init.ts; resolvePackageVersion now throws on failure so a broken dep
  setup surfaces loudly instead of silently publishing a stale default.
- CLI: detect Bun via npm_config_user_agent and use `bun install`.
- ESP32: trim oversized audit-false-positive comment in websocket_handler.c
  to match surrounding one-line comment style.

+++ firmware/pico/src/hal.cpp
@@ -61,9 +61,10 @@ void hal_set_gpio(uint8_t pin, gpio_state_t state) {
     }
 }
 
-// Track which pins have been configured as inputs (bitmask is 32 bits — safe because
-// is_valid_gpio() rejects pin >= NUM_BANK0_GPIOS before any shift).
-static uint32_t gpio_input_configured_mask = 0;
+// Track which pins have been configured as inputs. 64-bit mask covers both
+// RP2040 (NUM_BANK0_GPIOS=30) and RP2350 (NUM_BANK0_GPIOS=48); is_valid_gpio()
+// rejects pins outside that range before any shift.
+static uint64_t gpio_input_configured_mask = 0;
 
 bool hal_get_gpio_digital(uint8_t pin) {
     if (!is_valid_gpio(pin)) {
@@ -75,10 +76,10 @@ bool hal_get_gpio_digital(uint8_t pin) {
         return false;
     }
     // Only initialize the pin if not already configured as input
-    if (!(gpio_input_configured_mask & (1u << pin))) {
+    if (!(gpio_input_configured_mask & (1ull << pin))) {
         gpio_init(pin);
         gpio_set_dir(pin, GPIO_IN);
-        gpio_input_configured_mask |= (1u << pin);
+        gpio_input_configured_mask |= (1ull << pin);
     }
     return gpio_get(pin);
 }
@@ -96,7 +97,7 @@ void hal_configure_gpio_input(uint8_t pin, gpio_pull_t pull) {
     gpio_set_dir(pin, GPIO_IN);
 
     // Mark this pin as configured for input reads
-    gpio_input_configured_mask |= (1u << pin);
+    gpio_input_configured_mask |= (1ull << pin);
 
     // Configure pull-up or pull-down
     if (pull == GPIO_PULL_UP) {
